Refactor CharacterCount to CharacterCountMessage
